backend/requirements.txt
fastapi
uvicorn
google-generativeai
faiss-cpu
numpy
PyMuPDF  # for standard PDF text extraction
pytesseract
Pillow
python-multipart
python-jose
passlib
pydantic
pydantic-settings
python-dotenv
